Visual polishing

+++ src/lib/holidays.ts
@@ -1,6 +1,6 @@
 export function getFormattedHolidayDate(dateString: string): string {
   const date = new Date(dateString);
-  return date.toLocaleDateString("en-GB", { day: "numeric", month: "long" });
+  return date.toLocaleDateString("en-GB", { day: "numeric", month: "short" });
 }
 export function getFormattedDayOfWeek(dateString: string): string {
   const date = new Date(dateString);
